refactor: use common-dto in server
