+++ pom.xml
@@ -29,6 +29,7 @@
         <java.version>11</java.version>
         <lombok-mapstruct-binding.version>0.2.0</lombok-mapstruct-binding.version>
         <lombok.version>1.18.26</lombok.version>
+        <shareit-common-dto.version>0.0.1-SNAPSHOT</shareit-common-dto.version>
         <maven-checkstyle-plugin.version>3.1.2</maven-checkstyle-plugin.version>
         <maven-compiler-plugin.version>3.11.0</maven-compiler-plugin.version>
         <spotbugs-maven-plugin.version>4.7.0.0</spotbugs-maven-plugin.version>
