+++ server/pom.xml
@@ -56,6 +56,12 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
+        <dependency>
+            <groupId>ru.practicum</groupId>
+            <artifactId>shareit-common-dto</artifactId>
+            <version>${shareit-common-dto.version}</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 
     <build>
