--- server/src/main/java/ru/practicum/shareit/booking/dto/response/BookingDto.java
@@ -1,40 +1,0 @@
-package ru.practicum.shareit.booking.dto.response;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-import ru.practicum.shareit.booking.model.BookingStatus;
-
-import java.time.LocalDateTime;
-
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Setter
-@Getter
-@ToString
-@NoArgsConstructor
-public class BookingDto {
-    Long id;
-    LocalDateTime start;
-    LocalDateTime end;
-    BookingStatus status;
-    UserDto booker;
-    ItemDto item;
-
-    @FieldDefaults(level = AccessLevel.PRIVATE)
-    @Setter
-    @Getter
-    @ToString
-    @NoArgsConstructor
-    public static class ItemDto {
-        Long id;
-        String name;
-    }
-
-    @FieldDefaults(level = AccessLevel.PRIVATE)
-    @Setter
-    @Getter
-    @ToString
-    @NoArgsConstructor
-    public static class UserDto {
-        Long id;
-    }
-}
