--- server/src/main/java/ru/practicum/shareit/item/dto/request/comment/CommentCreationDto.java
@@ -1,13 +1,0 @@
-package ru.practicum.shareit.item.dto.request.comment;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@FieldDefaults(level = AccessLevel.PUBLIC)
-@Setter
-@Getter
-@ToString
-@NoArgsConstructor
-public class CommentCreationDto {
-    String text;
-}
