--- server/src/main/java/ru/practicum/shareit/item/dto/response/ItemDto.java
@@ -1,17 +1,0 @@
-package ru.practicum.shareit.item.dto.response;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-public class ItemDto {
-    Long id;
-    String name;
-    String description;
-    Boolean available;
-    Long requestId;
-}
