--- server/src/main/java/ru/practicum/shareit/item/dto/response/comment/CommentDto.java
@@ -1,18 +1,0 @@
-package ru.practicum.shareit.item.dto.response.comment;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-import java.time.LocalDateTime;
-
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Setter
-@Getter
-@ToString
-@NoArgsConstructor
-public class CommentDto {
-    Long id;
-    String text;
-    String authorName;
-    LocalDateTime created;
-}
