+++ server/src/main/java/ru/practicum/shareit/item/mapper/CommentDtoMapper.java
@@ -2,8 +2,8 @@ package ru.practicum.shareit.item.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import ru.practicum.shareit.item.dto.request.comment.CommentCreationDto;
-import ru.practicum.shareit.item.dto.response.comment.CommentDto;
+import ru.practicum.shareit.commondto.item.request.comment.CommentCreationDto;
+import ru.practicum.shareit.commondto.item.response.comment.CommentDto;
 import ru.practicum.shareit.item.model.Comment;
 import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.user.model.User;
