+++ server/src/main/java/ru/practicum/shareit/item/service/comment/CommentService.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.item.service.comment;
 
-import ru.practicum.shareit.item.dto.request.comment.CommentCreationDto;
-import ru.practicum.shareit.item.dto.response.comment.CommentDto;
+import ru.practicum.shareit.commondto.item.request.comment.CommentCreationDto;
+import ru.practicum.shareit.commondto.item.response.comment.CommentDto;
 
 public interface CommentService {
     CommentDto addComment(CommentCreationDto commentDto, Long authorId, Long itemId);
