+++ server/src/main/java/ru/practicum/shareit/item/service/comment/CommentServiceImpl.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.shareit.booking.model.BookingStatus;
 import ru.practicum.shareit.booking.repository.BookingRepository;
-import ru.practicum.shareit.item.dto.request.comment.CommentCreationDto;
-import ru.practicum.shareit.item.dto.response.comment.CommentDto;
+import ru.practicum.shareit.commondto.item.request.comment.CommentCreationDto;
+import ru.practicum.shareit.commondto.item.response.comment.CommentDto;
 import ru.practicum.shareit.item.exception.comment.CommentNotAllowedException;
 import ru.practicum.shareit.item.mapper.CommentDtoMapper;
 import ru.practicum.shareit.item.repository.ItemRepository;
