--- server/src/main/java/ru/practicum/shareit/itemrequest/dto/request/ItemRequestCreationDto.java
@@ -1,13 +1,0 @@
-package ru.practicum.shareit.itemrequest.dto.request;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Setter
-@Getter
-@ToString
-@NoArgsConstructor
-public class ItemRequestCreationDto {
-    String description;
-}
