--- server/src/main/java/ru/practicum/shareit/itemrequest/dto/response/ItemRequestDto.java
@@ -1,32 +1,0 @@
-package ru.practicum.shareit.itemrequest.dto.response;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Setter
-@Getter
-@ToString
-@NoArgsConstructor
-public class ItemRequestDto {
-    Long id;
-    String description;
-    LocalDateTime created;
-    List<ItemDto> items;
-
-    @FieldDefaults(level = AccessLevel.PRIVATE)
-    @Setter
-    @Getter
-    @ToString
-    @NoArgsConstructor
-    public static class ItemDto {
-        Long id;
-        String name;
-        String description;
-        Boolean available;
-        Long requestId;
-    }
-}
