--- server/src/main/java/ru/practicum/shareit/user/dto/request/UserCreationDto.java
@@ -1,14 +1,0 @@
-package ru.practicum.shareit.user.dto.request;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-public class UserCreationDto {
-    String name;
-    String email;
-}
