--- server/src/main/java/ru/practicum/shareit/user/dto/response/UserDto.java
@@ -1,15 +1,0 @@
-package ru.practicum.shareit.user.dto.response;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-public class UserDto {
-    Long id;
-    String name;
-    String email;
-}
